feat: like and unlike in one endpoint

+++ src/server/api/routers/like.ts
@@ -5,31 +5,41 @@ export const likeRouter = createTRPCRouter({
   like: protectedProcedure
     .input(z.object({ tweetId: z.string() }))
     .mutation(({ ctx, input }) => {
-      return ctx.prisma.like.create({
-        data: {
-          tweet: {
-            connect: {
-              id: input.tweetId,
+      return ctx.prisma.like
+        .findUnique({
+          where: {
+            userIdTweetId: {
+              userId: ctx.session.user.id,
+              tweetId: input.tweetId,
             },
           },
-          user: {
-            connect: {
-              id: ctx.session.user.id,
-            },
-          },
-        },
-      });
-    }),
-  unlike: protectedProcedure
-    .input(z.object({ tweetId: z.string() }))
-    .mutation(({ ctx, input }) => {
-      return ctx.prisma.like.delete({
-        where: {
-          userIdTweetId: {
-            userId: ctx.session.user.id,
-            tweetId: input.tweetId,
-          },
-        },
-      });
+        })
+        .then((like) => {
+          if (like) {
+            return ctx.prisma.like.delete({
+              where: {
+                userIdTweetId: {
+                  userId: ctx.session.user.id,
+                  tweetId: input.tweetId,
+                },
+              },
+            });
+          } else {
+            return ctx.prisma.like.create({
+              data: {
+                tweet: {
+                  connect: {
+                    id: input.tweetId,
+                  },
+                },
+                user: {
+                  connect: {
+                    id: ctx.session.user.id,
+                  },
+                },
+              },
+            });
+          }
+        });
     }),
 });
